+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Default Values - Fix the prefilled value in creation/update forms on Date fields having Sequelize.NOW default value.
+- Default Values - Make the record creation possible if some fields of the model have a default value function.
 
 ## RELEASE 2.5.8 - 2018-03-26
 ### Changed